ci(e2e): fix unused imports, split custom-env e2e, recalibrate with floor
